Move CacheGuardian to OapCache file and clean up the code in Guava cache as to the cache instance
